updating blueprints folder structure, refactoring app.py and models.py
