+++ Unit-02/01-blueprints/manage.py
@@ -1,12 +1,10 @@
-from app import app, db
+from project import app, db
 from flask_script import Manager
 from flask_migrate import Migrate, MigrateCommand
 
 migrate = Migrate(app, db)
 
 manager = Manager(app)
-
-
 manager.add_command('db', MigrateCommand) # python manage.py db
 
 if __name__ == '__main__':
