+++ Unit-02/01-blueprints/project/messages/views.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+
+messages_blueprint = Blueprint(
+    'messages',
+    __name__,
+    template_folder='templates'
+)
