+++ Unit-02/01-blueprints/project/users/views.py
@@ -1,0 +1,8 @@
+from flask import Blueprint
+from project.users.models import User
+
+users_blueprint = BLueprint(
+    'users',
+    __name__'
+    template_folder='templates'
+)
